Patch from Alessandro Premoli fixes all the warnings of scst on 64-bit platforms
git-svn-id: http://svn.code.sf.net/p/scst/svn/trunk@138 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ scst/src/dev_handlers/scst_dev_handler.h
@@ -42,7 +42,7 @@ static int dev_handler_log_info_show(struct seq_file *seq, void *v)
 	return res;
 }
 
-static int scst_dev_handler_proc_log_entry_write(struct file *file,
+static ssize_t scst_dev_handler_proc_log_entry_write(struct file *file,
 	const char __user *buf, size_t length, loff_t *off)
 {
 	int res = 0;
